refactor(router): share the PD KV-connector vocabulary across transports
The connector-mode derivation (Nixl/Mooncake/Passthrough), the NIXL
prefill tag, the Mooncake param synthesis, and the DP-aware engine-id
resolution lived inside the gRPC request-execution stage. The HTTP PD
router needs the identical vocabulary to speak sequential vLLM
disaggregation, so move it to routers/common/kv_transfer.rs and hoist
the per-worker derivation into connector_mode_for_worker. No behavior
change.
