+++ model_gateway/src/routers/grpc/common/stages/request_execution.rs
@@ -11,6 +11,10 @@ use super::PipelineStage;
 use crate::{
     observability::metrics::{metrics_labels, Metrics},
     routers::{
+        common::kv_transfer::{
+            connector_mode_for_worker, mooncake_decode_params, mooncake_prefill_params,
+            KvConnectorMode, NIXL_PREFILL_KV_PARAMS,
+        },
         error,
         grpc::{
             common::stages::encode::EncodeDispatchPlan,
@@ -25,9 +29,7 @@ use crate::{
             utils::tonic_ext::{TonicResultExt, TonicStatusExt},
         },
     },
-    worker::{
-        ConnectionModeExt, RuntimeType, DEFAULT_BOOTSTRAP_PORT, MOONCAKE_CONNECTOR, NIXL_CONNECTOR,
-    },
+    worker::{ConnectionModeExt, RuntimeType},
 };
 
 type StreamResult = Result<ProtoStream, tonic::Status>;
@@ -46,94 +48,6 @@ fn pd_leg_labels(workers: &WorkerSelection) -> (&'static str, &'static str) {
             (label, label)
         }
     }
-}
-
-/// KV-transfer params tagged onto the NIXL prefill leg so the engine pins its
-/// KV blocks and returns the handoff params for the decode worker.
-const NIXL_PREFILL_KV_PARAMS: &str = r#"{"do_remote_decode":true,"do_remote_prefill":false}"#;
-
-/// PD KV-transfer behavior derived from prefill worker metadata.
-#[derive(Debug, Clone, PartialEq)]
-enum KvConnectorMode {
-    /// MooncakeConnector: mint a transfer_id, tag both legs, synthesize decode
-    /// params from worker metadata; legacy host/port injection when the
-    /// servicer predates kv_engine_id reporting (or DP runs without a pinned rank).
-    Mooncake {
-        host: String,
-        port: u32,
-        engine_id: Option<String>,
-    },
-    /// NixlConnector: tag prefill with do_remote_decode, relay returned params to decode.
-    Nixl,
-    /// Unknown/absent connector: relay returned params opportunistically.
-    Passthrough,
-}
-
-impl KvConnectorMode {
-    fn metrics_label(&self) -> &'static str {
-        match self {
-            Self::Mooncake { .. } => metrics_labels::KV_CONNECTOR_MOONCAKE,
-            Self::Nixl => metrics_labels::KV_CONNECTOR_NIXL,
-            Self::Passthrough => metrics_labels::KV_CONNECTOR_PASSTHROUGH,
-        }
-    }
-}
-
-fn kv_connector_mode(
-    kv_connector: Option<&str>,
-    bootstrap_host: &str,
-    bootstrap_port: Option<u16>,
-    kv_engine_id: Option<&str>,
-) -> KvConnectorMode {
-    match kv_connector {
-        Some(MOONCAKE_CONNECTOR) => KvConnectorMode::Mooncake {
-            host: bootstrap_host.to_string(),
-            port: u32::from(bootstrap_port.unwrap_or(DEFAULT_BOOTSTRAP_PORT)),
-            // Empty means unknown (forces the legacy fallback)
-            engine_id: kv_engine_id.filter(|s| !s.is_empty()).map(str::to_string),
-        },
-        Some(NIXL_CONNECTOR) => KvConnectorMode::Nixl,
-        _ => KvConnectorMode::Passthrough,
-    }
-}
-
-/// Connector id of the engine core serving the prefill leg. With DP the cores
-/// suffix the configured id as `{base}_dp{rank}`, so minting needs a pinned
-/// rank; unpinned DP>1 yields None (no mint — decode recomputes locally).
-fn effective_kv_engine_id(
-    base: Option<&str>,
-    dp_size: Option<usize>,
-    dp_rank: Option<usize>,
-) -> Option<String> {
-    let base = base.filter(|s| !s.is_empty())?;
-    if dp_size.unwrap_or(1) > 1 {
-        dp_rank.map(|rank| format!("{base}_dp{rank}"))
-    } else {
-        Some(base.to_string())
-    }
-}
-
-/// Prefill-leg params for Mooncake: the engine pins blocks under the minted id.
-fn mooncake_prefill_params(transfer_id: &str) -> String {
-    serde_json::json!({
-        "do_remote_decode": true,
-        "do_remote_prefill": false,
-        "transfer_id": transfer_id,
-    })
-    .to_string()
-}
-
-/// Decode-leg params for Mooncake, synthesized from prefill worker metadata
-/// (the engine returns nothing to relay; the connector is push-based).
-fn mooncake_decode_params(transfer_id: &str, engine_id: &str, host: &str, port: u32) -> String {
-    serde_json::json!({
-        "do_remote_decode": false,
-        "do_remote_prefill": true,
-        "transfer_id": transfer_id,
-        "remote_engine_id": engine_id,
-        "remote_bootstrap_addr": format!("http://{host}:{port}"),
-    })
-    .to_string()
 }
 
 /// Request execution stage: execute the plan produced by request building.
@@ -588,22 +502,7 @@ impl RequestExecutionStage {
 
         let mode = workers
             .prefill_worker()
-            .map(|w| {
-                let meta = w.metadata();
-                // Discovered dp_size matters even without --dp-aware expansion:
-                // a DP>1 engine behind an unexpanded worker must not be minted for
-                let dp_size = w
-                    .dp_size()
-                    .or_else(|| meta.spec.labels.get("dp_size").and_then(|s| s.parse().ok()));
-                let engine_id =
-                    effective_kv_engine_id(meta.spec.kv_engine_id.as_deref(), dp_size, w.dp_rank());
-                kv_connector_mode(
-                    meta.spec.kv_connector.as_deref(),
-                    &meta.spec.bootstrap_host,
-                    meta.spec.bootstrap_port,
-                    engine_id.as_deref(),
-                )
-            })
+            .map(|w| connector_mode_for_worker(w.as_ref()))
             .unwrap_or(KvConnectorMode::Passthrough);
 
         // Recorded on the success path (after decode established) so failed
@@ -879,70 +778,6 @@ mod tests {
     }
 
     #[test]
-    fn kv_connector_mode_mooncake_uses_bootstrap_metadata() {
-        let mode = kv_connector_mode(
-            Some(MOONCAKE_CONNECTOR),
-            "prefill-host",
-            Some(9090),
-            Some("engine-1"),
-        );
-        assert_eq!(
-            mode,
-            KvConnectorMode::Mooncake {
-                host: "prefill-host".to_string(),
-                port: 9090,
-                engine_id: Some("engine-1".to_string()),
-            }
-        );
-    }
-
-    #[test]
-    fn kv_connector_mode_mooncake_defaults_port_and_tolerates_missing_engine_id() {
-        let mode = kv_connector_mode(Some(MOONCAKE_CONNECTOR), "prefill-host", None, None);
-        assert_eq!(
-            mode,
-            KvConnectorMode::Mooncake {
-                host: "prefill-host".to_string(),
-                port: u32::from(DEFAULT_BOOTSTRAP_PORT),
-                engine_id: None,
-            }
-        );
-    }
-
-    #[test]
-    fn kv_connector_mode_mooncake_empty_engine_id_means_legacy() {
-        let mode = kv_connector_mode(Some(MOONCAKE_CONNECTOR), "host", Some(9090), Some(""));
-        assert_eq!(
-            mode,
-            KvConnectorMode::Mooncake {
-                host: "host".to_string(),
-                port: 9090,
-                engine_id: None,
-            }
-        );
-    }
-
-    #[test]
-    fn kv_connector_mode_nixl() {
-        assert_eq!(
-            kv_connector_mode(Some(NIXL_CONNECTOR), "ignored", Some(9090), None),
-            KvConnectorMode::Nixl
-        );
-    }
-
-    #[test]
-    fn kv_connector_mode_unknown_or_missing_is_passthrough() {
-        assert_eq!(
-            kv_connector_mode(Some("LMCacheConnector"), "host", None, None),
-            KvConnectorMode::Passthrough
-        );
-        assert_eq!(
-            kv_connector_mode(None, "host", None, None),
-            KvConnectorMode::Passthrough
-        );
-    }
-
-    #[test]
     fn mooncake_prefill_params_carry_transfer_id() {
         let value: serde_json::Value =
             serde_json::from_str(&mooncake_prefill_params("xfer-abc")).unwrap();
@@ -1116,40 +951,5 @@ mod tests {
 
         let unset = ProtoGenerateComplete::Vllm(vllm::GenerateComplete::default());
         assert_eq!(unset.kv_transfer_params_json(), None);
-    }
-
-    #[test]
-    fn effective_engine_id_passthrough_when_no_dp() {
-        assert_eq!(
-            effective_kv_engine_id(Some("eng"), None, None).as_deref(),
-            Some("eng")
-        );
-        assert_eq!(
-            effective_kv_engine_id(Some("eng"), Some(1), None).as_deref(),
-            Some("eng")
-        );
-    }
-
-    #[test]
-    fn effective_engine_id_suffixes_pinned_dp_rank() {
-        assert_eq!(
-            effective_kv_engine_id(Some("eng"), Some(2), Some(1)).as_deref(),
-            Some("eng_dp1")
-        );
-        assert_eq!(
-            effective_kv_engine_id(Some("eng"), Some(2), Some(0)).as_deref(),
-            Some("eng_dp0")
-        );
-    }
-
-    #[test]
-    fn effective_engine_id_none_for_unpinned_dp() {
-        assert_eq!(effective_kv_engine_id(Some("eng"), Some(2), None), None);
-    }
-
-    #[test]
-    fn effective_engine_id_none_for_missing_or_empty_base() {
-        assert_eq!(effective_kv_engine_id(None, Some(2), Some(0)), None);
-        assert_eq!(effective_kv_engine_id(Some(""), None, None), None);
     }
 }
